requirements.txt
3to2==1.1.1
appdirs==1.4.3
applib==1.2
certifi==2019.9.11
chardet==3.0.4
cmdln==2.0.0
CodeIntel==2.0.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.5
django-allauth==0.40.0
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-mathfilters==0.4.0
django-tinymce4-lite==1.7.5
esprima==4.0.1
gunicorn==19.9.0
idna==2.8
Inflector==3.0.0
jsmin==2.2.2
libclang-py3==3.9.0
oauthlib==3.1.0
Pillow==6.1.0
psycopg2==2.8.3
pyenchant==2.0.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.5
whitenoise==4.1.3
zope.cachedescriptors==4.3.1
